basic broadcasing, turned bullet to int, basic HP

+++ Assets/Scripts/shooting scripts/BasicEnemyController.cs
@@ -7,7 +7,7 @@ public class BasicEnemyController : MonoBehaviour
     [SerializeField] float movementSpeed = 5f;
     [SerializeField] float shootInterval = 0.5f;
     [SerializeField] float bulletSpeed = 10f;
-    [SerializeField] float damage = 1f;
+    [SerializeField] int damage = 1;
     GameObject player;
     Rigidbody2D enemyBody;
     Vector3 finalDirection;
